cmd: initial refactor for multiple clis
