+++ cmd/buckd/main.go
@@ -15,172 +15,175 @@ import (
 )
 
 const (
-	daemonName = "bucketsd"
+	daemonName = "buckd"
 )
 
 var (
 	log = logging.Logger(daemonName)
 
-	configFile  string
-	configDir   = "." + daemonName
-	configViper = viper.New()
-
-	flags = map[string]cmd.Flag{
-		"repo": {
-			Key:      "repo",
-			DefValue: "${HOME}/" + configDir + "/repo",
+	config = cmd.Config{
+		Viper: viper.New(),
+		Dir:   "." + daemonName,
+		Name:  "config",
+		Flags: map[string]cmd.Flag{
+			"repo": {
+				Key:      "repo",
+				DefValue: "${HOME}/." + daemonName + "/repo",
+			},
+			"debug": {
+				Key:      "log.debug",
+				DefValue: false,
+			},
+			"logFile": {
+				Key:      "log.file",
+				DefValue: "${HOME}/." + daemonName + "/log",
+			},
+			"addrApi": {
+				Key:      "addr.api",
+				DefValue: "/ip4/127.0.0.1/tcp/3006",
+			},
+			"addrApiProxy": {
+				Key:      "addr.api_proxy",
+				DefValue: "/ip4/127.0.0.1/tcp/3007",
+			},
+			"addrThreadsHost": {
+				Key:      "addr.threads.host",
+				DefValue: "/ip4/0.0.0.0/tcp/4006",
+			},
+			"addrGatewayHost": {
+				Key:      "addr.gateway.host",
+				DefValue: "/ip4/127.0.0.1/tcp/8006",
+			},
+			"addrGatewayUrl": {
+				Key:      "addr.gateway.url",
+				DefValue: "http://127.0.0.1:8006",
+			},
+			"addrIpfsApi": {
+				Key:      "addr.ipfs.api",
+				DefValue: "/ip4/127.0.0.1/tcp/5001",
+			},
+			"addrFilecoinApi": {
+				Key:      "addr.filecoin.api",
+				DefValue: "",
+			},
+			"addrMongoUri": {
+				Key:      "addr.mongo_uri",
+				DefValue: "mongodb://127.0.0.1:27017",
+			},
+			"gatewaySubdomains": {
+				Key:      "gateway.subdomains",
+				DefValue: false,
+			},
+			"dnsDomain": {
+				Key:      "dns.domain",
+				DefValue: "",
+			},
+			"dnsZoneID": {
+				Key:      "dns.zone_id",
+				DefValue: "",
+			},
+			"dnsToken": {
+				Key:      "dns.token",
+				DefValue: "",
+			},
 		},
-		"debug": {
-			Key:      "log.debug",
-			DefValue: false,
-		},
-		"logFile": {
-			Key:      "log.file",
-			DefValue: "${HOME}/" + configDir + "/log",
-		},
-		"addrApi": {
-			Key:      "addr.api",
-			DefValue: "/ip4/127.0.0.1/tcp/3006",
-		},
-		"addrApiProxy": {
-			Key:      "addr.api_proxy",
-			DefValue: "/ip4/127.0.0.1/tcp/3007",
-		},
-		"addrThreadsHost": {
-			Key:      "addr.threads.host",
-			DefValue: "/ip4/0.0.0.0/tcp/4006",
-		},
-		"addrGatewayHost": {
-			Key:      "addr.gateway.host",
-			DefValue: "/ip4/127.0.0.1/tcp/8006",
-		},
-		"addrGatewayUrl": {
-			Key:      "addr.gateway.url",
-			DefValue: "http://127.0.0.1:8006",
-		},
-		"addrIpfsApi": {
-			Key:      "addr.ipfs.api",
-			DefValue: "/ip4/127.0.0.1/tcp/5001",
-		},
-		"addrFilecoinApi": {
-			Key:      "addr.filecoin.api",
-			DefValue: "",
-		},
-		"addrMongoUri": {
-			Key:      "addr.mongo_uri",
-			DefValue: "mongodb://127.0.0.1:27017",
-		},
-		"gatewaySubdomains": {
-			Key:      "gateway.subdomains",
-			DefValue: false,
-		},
-		"dnsDomain": {
-			Key:      "dns.domain",
-			DefValue: "",
-		},
-		"dnsZoneID": {
-			Key:      "dns.zone_id",
-			DefValue: "",
-		},
-		"dnsToken": {
-			Key:      "dns.token",
-			DefValue: "",
-		},
+		EnvPre: "BUCK",
+		Global: true,
 	}
 )
 
 func init() {
-	cobra.OnInitialize(cmd.InitConfig(configViper, configFile, configDir, "config", true))
-	cmd.InitConfigCmd(rootCmd, configViper, configDir)
+	cobra.OnInitialize(cmd.InitConfig(config))
+	cmd.InitConfigCmd(rootCmd, config.Viper, config.Dir)
 
 	rootCmd.PersistentFlags().StringVar(
-		&configFile,
+		&config.File,
 		"config",
 		"",
-		"Config file (default ${HOME}/"+configDir+"/config.yml)")
+		"Config file (default ${HOME}/"+config.Dir+"/"+config.Name+".yml)")
 
 	rootCmd.PersistentFlags().StringP(
 		"repo",
 		"r",
-		flags["repo"].DefValue.(string),
+		config.Flags["repo"].DefValue.(string),
 		"Path to repository")
 
 	rootCmd.PersistentFlags().BoolP(
 		"debug",
 		"d",
-		flags["debug"].DefValue.(bool),
+		config.Flags["debug"].DefValue.(bool),
 		"Enable debug logging")
 
 	rootCmd.PersistentFlags().String(
 		"logFile",
-		flags["logFile"].DefValue.(string),
+		config.Flags["logFile"].DefValue.(string),
 		"Write logs to file")
 
 	rootCmd.PersistentFlags().String(
 		"addrApi",
-		flags["addrApi"].DefValue.(string),
-		"Textile API listen address")
+		config.Flags["addrApi"].DefValue.(string),
+		"Hub API listen address")
 
 	rootCmd.PersistentFlags().String(
 		"addrApiProxy",
-		flags["addrApiProxy"].DefValue.(string),
-		"Textile API proxy listen address")
+		config.Flags["addrApiProxy"].DefValue.(string),
+		"Hub API proxy listen address")
 
 	rootCmd.PersistentFlags().String(
 		"addrThreadsHost",
-		flags["addrThreadsHost"].DefValue.(string),
+		config.Flags["addrThreadsHost"].DefValue.(string),
 		"Threads peer host listen address")
 
 	rootCmd.PersistentFlags().String(
 		"addrIpfsApi",
-		flags["addrIpfsApi"].DefValue.(string),
+		config.Flags["addrIpfsApi"].DefValue.(string),
 		"IPFS API address")
 
 	// Gateway settings
 	rootCmd.PersistentFlags().String(
 		"addrGatewayHost",
-		flags["addrGatewayHost"].DefValue.(string),
+		config.Flags["addrGatewayHost"].DefValue.(string),
 		"Local gateway host address")
 	rootCmd.PersistentFlags().String(
 		"addrGatewayUrl",
-		flags["addrGatewayUrl"].DefValue.(string),
+		config.Flags["addrGatewayUrl"].DefValue.(string),
 		"Public gateway address")
 
 	// Filecoin settings
 	rootCmd.PersistentFlags().String(
 		"addrFilecoinApi",
-		flags["addrFilecoinApi"].DefValue.(string),
+		config.Flags["addrFilecoinApi"].DefValue.(string),
 		"Filecoin gRPC API address")
 
 	// Mongo settings
 	rootCmd.PersistentFlags().String(
 		"addrMongoUri",
-		flags["addrMongoUri"].DefValue.(string),
+		config.Flags["addrMongoUri"].DefValue.(string),
 		"MongoDB connection URI")
 
 	// Gateway settings
 	rootCmd.PersistentFlags().Bool(
 		"gatewaySubdomains",
-		flags["gatewaySubdomains"].DefValue.(bool),
+		config.Flags["gatewaySubdomains"].DefValue.(bool),
 		"Enable gateway namespace redirects to subdomains")
 
 	// Cloudflare settings
 	rootCmd.PersistentFlags().String(
 		"dnsDomain",
-		flags["dnsDomain"].DefValue.(string),
+		config.Flags["dnsDomain"].DefValue.(string),
 		"Root domain for bucket subdomains")
 
 	rootCmd.PersistentFlags().String(
 		"dnsZoneID",
-		flags["dnsZoneID"].DefValue.(string),
+		config.Flags["dnsZoneID"].DefValue.(string),
 		"Cloudflare ZoneID for dnsDomain")
 
 	rootCmd.PersistentFlags().String(
 		"dnsToken",
-		flags["dnsDomain"].DefValue.(string),
+		config.Flags["dnsDomain"].DefValue.(string),
 		"Cloudflare API Token for dnsDomain")
 
-	if err := cmd.BindFlags(configViper, rootCmd, flags); err != nil {
+	if err := cmd.BindFlags(config.Viper, rootCmd, config.Flags); err != nil {
 		log.Fatal(err)
 	}
 }
@@ -196,10 +199,10 @@ var rootCmd = &cobra.Command{
 	Short: "Buckets daemon",
 	Long:  `The Buckets daemon.`,
 	PersistentPreRun: func(c *cobra.Command, args []string) {
-		configViper.SetConfigType("yaml")
-		cmd.ExpandConfigVars(configViper, flags)
+		config.Viper.SetConfigType("yaml")
+		cmd.ExpandConfigVars(config.Viper, config.Flags)
 
-		if configViper.GetBool("log.debug") {
+		if config.Viper.GetBool("log.debug") {
 			if err := util.SetLogLevels(map[string]logging.LogLevel{
 				daemonName: logging.LevelDebug,
 			}); err != nil {
@@ -208,32 +211,32 @@ var rootCmd = &cobra.Command{
 		}
 	},
 	Run: func(c *cobra.Command, args []string) {
-		settings, err := json.MarshalIndent(configViper.AllSettings(), "", "  ")
+		settings, err := json.MarshalIndent(config.Viper.AllSettings(), "", "  ")
 		if err != nil {
 			log.Fatal(err)
 		}
 		log.Debugf("loaded config: %s", string(settings))
 
-		addrApi := cmd.AddrFromStr(configViper.GetString("addr.api"))
-		addrApiProxy := cmd.AddrFromStr(configViper.GetString("addr.api_proxy"))
-		addrThreadsHost := cmd.AddrFromStr(configViper.GetString("addr.threads.host"))
-		addrIpfsApi := cmd.AddrFromStr(configViper.GetString("addr.ipfs.api"))
+		addrApi := cmd.AddrFromStr(config.Viper.GetString("addr.api"))
+		addrApiProxy := cmd.AddrFromStr(config.Viper.GetString("addr.api_proxy"))
+		addrThreadsHost := cmd.AddrFromStr(config.Viper.GetString("addr.threads.host"))
+		addrIpfsApi := cmd.AddrFromStr(config.Viper.GetString("addr.ipfs.api"))
 
-		addrGatewayHost := cmd.AddrFromStr(configViper.GetString("addr.gateway.host"))
-		addrGatewayUrl := configViper.GetString("addr.gateway.url")
+		addrGatewayHost := cmd.AddrFromStr(config.Viper.GetString("addr.gateway.host"))
+		addrGatewayUrl := config.Viper.GetString("addr.gateway.url")
 
 		var addrPowergateApi ma.Multiaddr
-		if str := configViper.GetString("addr.powergate.api"); str != "" {
+		if str := config.Viper.GetString("addr.powergate.api"); str != "" {
 			addrPowergateApi = cmd.AddrFromStr(str)
 		}
 
-		addrMongoUri := configViper.GetString("addr.mongo_uri")
+		addrMongoUri := config.Viper.GetString("addr.mongo_uri")
 
-		dnsDomain := configViper.GetString("dns.domain")
-		dnsZoneID := configViper.GetString("dns.zone_id")
-		dnsToken := configViper.GetString("dns.token")
+		dnsDomain := config.Viper.GetString("dns.domain")
+		dnsZoneID := config.Viper.GetString("dns.zone_id")
+		dnsToken := config.Viper.GetString("dns.token")
 
-		logFile := configViper.GetString("log.file")
+		logFile := config.Viper.GetString("log.file")
 		if logFile != "" {
 			util.SetupDefaultLoggingConfig(logFile)
 		}
@@ -241,7 +244,7 @@ var rootCmd = &cobra.Command{
 		ctx, cancel := context.WithCancel(context.Background())
 		defer cancel()
 		textile, err := core.NewTextile(ctx, core.Config{
-			RepoPath:         configViper.GetString("repo"),
+			RepoPath:         config.Viper.GetString("repo"),
 			AddrAPI:          addrApi,
 			AddrAPIProxy:     addrApiProxy,
 			AddrThreadsHost:  addrThreadsHost,
@@ -250,12 +253,12 @@ var rootCmd = &cobra.Command{
 			AddrGatewayURL:   addrGatewayUrl,
 			AddrPowergateAPI: addrPowergateApi,
 			AddrMongoURI:     addrMongoUri,
-			UseSubdomains:    configViper.GetBool("gateway.subdomains"),
+			UseSubdomains:    config.Viper.GetBool("gateway.subdomains"),
 			MongoName:        "buck",
 			DNSDomain:        dnsDomain,
 			DNSZoneID:        dnsZoneID,
 			DNSToken:         dnsToken,
-			Debug:            configViper.GetBool("log.debug"),
+			Debug:            config.Viper.GetBool("log.debug"),
 		})
 		if err != nil {
 			log.Fatal(err)
