+++ cmd/output.go
@@ -1,0 +1,76 @@
+package cmd
+
+import (
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/logrusorgru/aurora"
+	"github.com/olekukonko/tablewriter"
+	"google.golang.org/grpc/status"
+)
+
+func Message(format string, args ...interface{}) {
+	if format == "" {
+		return
+	}
+	fmt.Println(aurora.Sprintf(aurora.BrightBlack("> "+format), args...))
+}
+
+func Warn(format string, args ...interface{}) {
+	if format == "" {
+		return
+	}
+	fmt.Println(aurora.Sprintf(aurora.Yellow("! "+format), args...))
+}
+
+func Success(format string, args ...interface{}) {
+	fmt.Println(aurora.Sprintf(aurora.Cyan("> Success! %s"),
+		aurora.Sprintf(aurora.BrightBlack(format), args...)))
+}
+
+func End(format string, args ...interface{}) {
+	Message(format, args...)
+	os.Exit(0)
+}
+
+func Fatal(err error, args ...interface{}) {
+	var msg string
+	stat, ok := status.FromError(err)
+	if ok {
+		msg = stat.Message()
+	} else {
+		msg = err.Error()
+	}
+	words := strings.SplitN(msg, " ", 2)
+	words[0] = strings.Title(words[0])
+	msg = strings.Join(words, " ")
+	fmt.Println(aurora.Sprintf(aurora.Red("> Error! %s"),
+		aurora.Sprintf(aurora.BrightBlack(msg), args...)))
+	os.Exit(1)
+}
+
+func RenderTable(header []string, data [][]string) {
+	fmt.Println()
+	table := tablewriter.NewWriter(os.Stdout)
+	table.SetHeader(header)
+	table.SetAutoWrapText(false)
+	table.SetAutoFormatHeaders(true)
+	table.SetHeaderAlignment(tablewriter.ALIGN_LEFT)
+	table.SetAlignment(tablewriter.ALIGN_LEFT)
+	table.SetCenterSeparator("")
+	table.SetColumnSeparator("")
+	table.SetRowSeparator("")
+	table.SetHeaderLine(false)
+	table.SetBorder(false)
+	table.SetTablePadding("\t")
+	table.SetNoWhiteSpace(false)
+	headersColors := make([]tablewriter.Colors, len(data[0]))
+	for i := range headersColors {
+		headersColors[i] = tablewriter.Colors{tablewriter.FgHiBlackColor}
+	}
+	table.SetHeaderColor(headersColors...)
+	table.AppendBulk(data)
+	table.Render()
+	fmt.Println()
+}
