+++ email/templates.go
@@ -1,6 +1,6 @@
 package email
 
-const headerMsg = `Dear Textile developer,
+const headerMsg = `Dear developer,
 `
 
 const footerMsg = `
@@ -17,7 +17,7 @@ To complete the login process, follow the link below:
 ` + footerMsg
 
 const inviteMsg = headerMsg + `
-{{.From}} has invited you to the {{.Org}} organization on Textile.
+{{.From}} has invited you to the {{.Org}} organization on the Hub.
 
 To accept the invitation, follow the link below:
