+++ gateway/public/html/consent.gohtml
@@ -5,7 +5,7 @@
     </div>
     <div class="aligner-item">
         <p><b>{{.Email}}</b> is now a member of the <b>{{.Org}}</b> organization.</p>
-        <p>If you haven't already, <a href="https://github.com/textileio/textile/releases">download the Textile CLI</a> and initialize a new account. Check out <a href="https://github.com/textileio/textile">the docs</a> for more. You may now close this window!</p>
+        <p>If you haven't already, <a href="https://github.com/textileio/textile/releases">download the Hub CLI</a> and initialize a new account. Check out <a href="https://github.com/textileio/textile">the docs</a> for more. You may now close this window!</p>
     </div>
 </div>
 {{template "footer"}}
